Revert "Runtime Support for new parallel type (#1022)"
This reverts commit 6b523af940a9fccce7b0f291941005e3cbce4aae, which
causes a build error:

```
      FAILED: CMakeFiles/nvfuser_codegen.dir/csrc/executor.cpp.o
      /opt/cache/bin/sccache /usr/bin/c++ -DTORCH_CUDA_BUILD_MAIN_LIB -DUSE_C10D_GLOO -DUSE_C10D_NCCL -DUSE_DISTRIBUTED -DUSE_RPC -DUSE_TENSORPIPE -Dnvfuser_codegen_EXPORTS -I/opt/pytorch/nvfuser/cmake/../third_party/benchmark/include -I/opt/pytorch/nvfuser/third_party/gloo -I/opt/pytorch/nvfuser/csrc -I/opt/pytorch/nvfuser/build/include -I/opt/pytorch/nvfuser/third_party/flatbuffers/include -I/opt/pytorch/nvfuser/lib/dynamic_type/src -isystem /opt/pytorch/nvfuser/cmake/../third_party/googletest/googlemock/include -isystem /opt/pytorch/nvfuser/cmake/../third_party/googletest/googletest/include -isystem /opt/pytorch/nvfuser/cmake/../third_party/flatbuffers/include -isystem /usr/local/cuda/include -isystem /usr/local/lib/python3.10/site-packages/torch/include -isystem /usr/local/lib/python3.10/site-packages/torch/include/torch/csrc/api/include -Wno-psabi -D_GLIBCXX_USE_CXX11_ABI=0 -O3 -DNDEBUG -fPIC -Wall -Wno-unused-function -DTORCH_CUDA_BUILD_MAIN_LIB -Werror -D_GLIBCXX_USE_CXX11_ABI=0 -MD -MT CMakeFiles/nvfuser_codegen.dir/csrc/executor.cpp.o -MF CMakeFiles/nvfuser_codegen.dir/csrc/executor.cpp.o.d -o CMakeFiles/nvfuser_codegen.dir/csrc/executor.cpp.o -c /opt/pytorch/nvfuser/csrc/executor.cpp
      /opt/pytorch/nvfuser/csrc/executor.cpp: In member function ‘std::vector<at::Tensor> nvfuser::FusionExecutor::allocOutputSpace(const c10::ArrayRef<c10::IValue>&)’:
      /opt/pytorch/nvfuser/csrc/executor.cpp:1269:36: error: ‘InputOutputAliases’ is not a member of ‘nvfuser::executor_utils::caching’
       1269 |           executor_utils::caching::InputOutputAliases>(
            |                                    ^~~~~~~~~~~~~~~~~~
      /opt/pytorch/nvfuser/csrc/executor.cpp:1269:54: error: template argument 1 is invalid
       1269 |           executor_utils::caching::InputOutputAliases>(
            |                                                      ^
      /opt/pytorch/nvfuser/csrc/executor.cpp: In lambda function:
      /opt/pytorch/nvfuser/csrc/executor.cpp:1271:49: error: ‘InputOutputAlias’ was not declared in this scope
       1271 |             return std::make_unique<std::vector<InputOutputAlias>>(
            |                                                 ^~~~~~~~~~~~~~~~
      /opt/pytorch/nvfuser/csrc/executor.cpp:1271:49: error: template argument 1 is invalid
      /opt/pytorch/nvfuser/csrc/executor.cpp:1271:49: error: template argument 2 is invalid
      /opt/pytorch/nvfuser/csrc/executor.cpp:1272:26: error: ‘class nvfuser::Fusion’ has no member named ‘getOutputToInputAliasIndices’
       1272 |                 fusion_->getOutputToInputAliasIndices());
            |                          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

+++ csrc/multidevice/communicator.cpp
@@ -116,7 +116,7 @@ c10::intrusive_ptr<c10d::Backend> createBackend(
     CommunicatorBackend backend,
     ::c10::intrusive_ptr<c10d::Store> store,
     RankType rank,
-    int64_t size) {
+    uint64_t size) {
 #ifdef USE_C10D_NCCL
   if (backend == CommunicatorBackend::nccl) {
     auto pg_opts = c10::make_intrusive<::c10d::ProcessGroupNCCL::Options>();
@@ -198,7 +198,7 @@ c10::intrusive_ptr<c10d::Backend> Communicator::getBackendForTeam(
         backend_type_,
         c10::make_intrusive<c10d::PrefixStore>(team_key, store_),
         team_rank,
-        static_cast<int64_t>(team.size()));
+        team.size());
   }
   return backends_.at(team_key);
 }
